+++ src/main/java/com/cedarsoftware/io/Writers.java
@@ -108,7 +108,9 @@ public class Writers {
          */
         @SuppressWarnings("unchecked")
         public void writePrimitiveForm(Object o, Writer output, WriterContext context) throws IOException {
-            if (context.getWriteOptions().isAllowNanAndInfinity() || !isNanOrInfinity((T) o)) {
+            WriteOptions options = context.getWriteOptions();
+            boolean allowNanInfinity = options.isAllowNanAndInfinity() || options.isJson5InfinityNaN();
+            if (allowNanInfinity || !isNanOrInfinity((T) o)) {
                 output.write(o.toString());
             } else {
                 output.write("null");